// src/__generated__/AppRepositoryListComponent_repository.graphql.ts
export type AppRepositoryListComponent_repository$data = {
  readonly search: {
    readonly edges: ReadonlyArray<{
      readonly cursor: string;
      readonly node: {
        readonly description?: string | null;
        readonly name?: string;
        readonly stargazerCount?: number;
      } | null;
    } | null> | null;
    readonly pageInfo: {
      readonly endCursor: string | null;
      readonly hasNextPage: boolean;
    };
  };
  readonly " $fragmentType": "AppRepositoryListComponent_repository";
};
export type AppRepositoryListComponent_repository$key = {
  readonly " $data"?: AppRepositoryListComponent_repository$data;
  readonly " $fragmentSpreads": FragmentRefs<"AppRepositoryListComponent_repository">;
};